# docs/reference/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "description": "The JSON body of every error response.",
        "properties": {
          "error": {
            "title": "Error",
            "type": "string"
          },
          "issues": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issues"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "results": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Results"
          }
        },
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "ExpectedLink": {
        "additionalProperties": false,
        "description": "One declared expectation.",
        "properties": {
          "by_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "By State"
          },
          "direction": {
            "default": "out",
            "enum": [
              "out",
              "in",
              "either"
            ],
            "title": "Direction",
            "type": "string"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "min_count": {
            "default": 1,
            "minimum": 1.0,
            "title": "Min Count",
            "type": "integer"
          },
          "relation": {
            "title": "Relation",
            "type": "string"
          },
          "target_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Type"
          }
        },
        "required": [
          "relation"
        ],
        "title": "ExpectedLink",
        "type": "object"
      },
...
      "LinkCounts": {
        "properties": {
          "active_in": {
            "default": 0,
            "title": "Active In",
            "type": "integer"
          },
          "active_out": {
            "default": 0,
            "title": "Active Out",
            "type": "integer"
          },
          "broken": {
            "default": 0,
            "title": "Broken",
            "type": "integer"
          },
          "record_id": {
            "title": "Record Id",
            "type": "string"
          },
          "stale": {
            "default": 0,
            "title": "Stale",
            "type": "integer"
          },
          "suggested": {
            "default": 0,
            "title": "Suggested",
            "type": "integer"
          }
        },
        "required": [
          "record_id"
        ],
        "title": "LinkCounts",
        "type": "object"
      },
      "LinkTarget": {
        "description": "A record offered by the link picker.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key"
          },
          "link_total": {
            "title": "Link Total",
            "type": "integer"
          },
          "scope": {
            "title": "Scope",
            "type": "string"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "key",
          "type",
          "title",
          "status",
          "scope",
          "link_total"
        ],
        "title": "LinkTarget",
        "type": "object"
      },
      "LinkView": {
        "description": "One link as seen from one record.",
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "declined": {
            "title": "Declined",
            "type": "boolean"
          },
          "direction": {
            "enum": [
              "out",
              "in"
            ],
            "title": "Direction",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "link_id": {
            "title": "Link Id",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "other_id": {
            "title": "Other Id",
            "type": "string"
          },
          "other_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Other Key"
          },
          "other_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Other Status"
          },
          "other_title": {
            "title": "Other Title",
            "type": "string"
          },
          "other_type": {
            "title": "Other Type",
            "type": "string"
          },
          "other_voided": {
            "title": "Other Voided",
            "type": "boolean"
          },
          "pin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pin"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "relation": {
            "title": "Relation",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "verified_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified At"
          },
          "verified_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified By"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "link_id",
          "direction",
          "relation",
          "label",
          "other_id",
          "other_key",
          "other_title",
          "other_type",
          "other_status",
          "other_voided",
          "status",
          "pin",
          "note",
          "source",
          "confidence",
          "reason",
          "declined",
          "verified_by",
          "verified_at",
          "created_at",
          "version"
        ],
        "title": "LinkView",
        "type": "object"
      },
      "MissingLink": {
        "description": "An expectation that is not met: how many matching links exist and how many are needed.",
        "properties": {
          "expectation": {
            "$ref": "#/components/schemas/ExpectedLink"
          },
          "found": {
            "title": "Found",
            "type": "integer"
          },
          "needed": {
            "title": "Needed",
            "type": "integer"
          }
        },
        "required": [
          "expectation",
          "found",
          "needed"
        ],
        "title": "MissingLink",
        "type": "object"
      },
...
      "TraceNode": {
        "properties": {
          "children": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/TraceNode"
            },
            "title": "Children",
            "type": "array"
          },
          "depth": {
            "title": "Depth",
            "type": "integer"
          },
          "direction": {
            "anyOf": [
              {
                "enum": [
                  "out",
                  "in"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direction"
          },
          "key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "link_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Link Id"
          },
          "link_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Link Status"
          },
          "more": {
            "default": 0,
            "title": "More",
            "type": "integer"
          },
          "record_id": {
            "title": "Record Id",
            "type": "string"
          },
          "relation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relation"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "voided": {
            "title": "Voided",
            "type": "boolean"
          }
        },
        "required": [
          "record_id",
          "key",
          "title",
          "type",
          "status",
          "voided",
          "depth"
        ],
        "title": "TraceNode",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Static dev token (ADR-0005)",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Records, links, workflow, files and the change feed over HTTP. Phase 0 is a development server: identity is a static bearer token and every authenticated actor may do everything (ADR-0005). Errors are `{error, message, ...}` bodies; the `error` codes are stable.",
    "title": "Throughline API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/links/counts": {
      "get": {
        "description": "Link counts for each record id given (zeros for a record with no links).",
        "operationId": "get_link_counts",
        "parameters": [
          {
            "in": "query",
            "name": "record_id",
            "required": true,
            "schema": {
              "items": {
                "type": "string"
              },
              "maxItems": 500,
              "minItems": 1,
              "title": "Record Id",
              "type": "array"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "$ref": "#/components/schemas/LinkCounts"
                  },
                  "title": "Response Get Link Counts",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or unknown bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused by the authorisation hook or by a quarantine rule."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such resource."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Version conflict, state conflict or failed workflow guard."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The request is well formed but cannot be applied."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Link Counts",
        "tags": [
          "links"
        ]
      }
    },
    "/links/search": {
      "get": {
        "description": "Records the link picker may offer from this scope (its own and the company's).",
        "operationId": "search_linkable",
        "parameters": [
          {
            "in": "query",
            "name": "scope",
            "required": true,
            "schema": {
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "description": "Words that must occur in the key or title.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "description": "Words that must occur in the key or title.",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "record_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Record Type"
            }
          },
          {
            "in": "query",
            "name": "exclude_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Exclude Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LinkTarget"
                  },
                  "title": "Response Search Linkable",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or unknown bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused by the authorisation hook or by a quarantine rule."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such resource."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Version conflict, state conflict or failed workflow guard."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The request is well formed but cannot be applied."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Search Linkable",
        "tags": [
          "links"
        ]
      }
    },
...
    "/records/{record_id}/expected-links": {
      "get": {
        "description": "Expected links the record does not have yet.",
        "operationId": "get_expected_links",
        "parameters": [
          {
            "in": "path",
            "name": "record_id",
            "required": true,
            "schema": {
              "title": "Record Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MissingLink"
                  },
                  "title": "Response Get Expected Links",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or unknown bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused by the authorisation hook or by a quarantine rule."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such resource."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Version conflict, state conflict or failed workflow guard."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The request is well formed but cannot be applied."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Expected Links",
        "tags": [
          "links"
        ]
      }
    },
...
    "/records/{record_id}/links": {
      "get": {
        "description": "Both directions of the record's links, each with its label as read from this record.",
        "operationId": "get_record_links",
        "parameters": [
          {
            "in": "path",
            "name": "record_id",
            "required": true,
            "schema": {
              "title": "Record Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_retracted",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Retracted",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LinkView"
                  },
                  "title": "Response Get Record Links",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or unknown bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused by the authorisation hook or by a quarantine rule."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such resource."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Version conflict, state conflict or failed workflow guard."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The request is well formed but cannot be applied."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Record Links",
        "tags": [
          "links"
        ]
      }
    },
    "/records/{record_id}/trace": {
      "get": {
        "description": "The tree of records within `depth` link hops of the record (brief 7.5).",
        "operationId": "get_record_trace",
        "parameters": [
          {
            "in": "path",
            "name": "record_id",
            "required": true,
            "schema": {
              "title": "Record Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "depth",
            "required": false,
            "schema": {
              "default": 2,
              "maximum": 8,
              "minimum": 0,
              "title": "Depth",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "direction",
            "required": false,
            "schema": {
              "default": "both",
              "enum": [
                "out",
                "in",
                "both"
              ],
              "title": "Direction",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraceNode"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or unknown bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused by the authorisation hook or by a quarantine rule."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such resource."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Version conflict, state conflict or failed workflow guard."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The request is well formed but cannot be applied."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Record Trace",
        "tags": [
          "links"
        ]
      }
    },
...
  }
}